+++ packages/hume-app/server/lib/utils.js
@@ -10,30 +10,9 @@
 'use strict';
 
 const utils = require('@hume/utils');
-const monit = require('@hume/monit-express');
 
 const { name } = require('../../package.json');
 
-const log = utils.logger(name);
-
-utils.log = log;
-
-// To set the env var is the way we use to enable it.
-if (process.env.DB_ELASTIC) {
-  utils.monit = monit;
-}
-
-utils.error = async (msg, error, opts) => {
-  if (process.env.DB_ELASTIC) {
-    try {
-      await monit.error(msg, error, opts);
-      log.debug('Error properly reported to monit', error);
-    } catch (err) {
-      log.error('monit reporting error: ', err);
-    }
-  }
-
-  log.error(msg, error);
-};
+utils.log = utils.logger(name);
 
 module.exports = utils;